upscale int8 to int16 for numpy 2.0

+++ pocean/tests/test_utils.py
@@ -138,7 +138,7 @@ class TestUtils(unittest.TestCase):
             # to make sure it doesn't error
             b = ncd.createVariable('imabyte', 'b')
             b.valid_min = 0
-            b.valid_max = 600  # this is over a byte and thus invalid
+            b.valid_max = np.int16(600)  # this is over a byte and thus invalid
             b[:] = 3
             r = generic_masked(b[:], attrs=ncd.vatts(b.name))
             assert np.all(r.mask == False)  # noqa
